requirements.txt
requests==2.31.0
aiohttp==3.8.1
Pillow==8.4.0
matplotlib==3.8.2